[core, sim, sdk] Code fixes and review and refactor of NMF Services
Browsable version of the services is available under:
https://dmarszk.github.io/MOWebViewer4NMF/

Simulator and example application source base have been
adjusted accordingly.

General changes to the Platform Services:
* Changed structure and operation names to be more self-explanatory
* Removed Magnetometer service (duplicates AutonomousADCS functionality)
* Improved comments for all of the structures and operations
* Downgraded Double numbers to single (32-bit) floating point
  format (Float) in Platform service structures.
  iADCS API uses only float, and it is significantly
  faster than double on the embedded JRE.

Camera Service:
* Reordered fields in Camera::Picture composite, so that blob
  buffer is the last field in the message body
* Removed Camera::unsetStream -
  both enable/disable are now toggled through enableStream operation

SDR Service:
* Changed SDR Configuration members to match SEPP SDR API

Autonomous ADCS Service:
* Corrected quaternion field names and descriptions
* Added getAttitude request operation to AutonomousADCS
* Removed unsetAttitude (now triggered through setDesiredAttitude with null AttitudeMode)
* Decoupled attitude monitoring from attitude control capability
* Refactored ADCS interfaces to expose full attitude and actuator
  telemetry independently of the chosen control algorithm.
* Removed use of the COM in the ADCS service

Power Control Service:
* Reworked PowerControl to allow control of the devices by device type (i.e. OBC, SBAND, CAMERA, OPTRX)

Optical Data Receiver Service:
* Reworked the service interface to accomodate the fact, that the
  Optical RX on OPS-SAT has no streaming capability, and
  the functionality covers exclusive modes of either
  "recording samples to memory" or "readout from memory".

General changes to the Software Management services:
* Removed services: Memory Management, Software Image, Processes
* Introduced Command Executor Service, allowing to invoke arbitrary
  commands in the service provider host shell.

+++ mission/simulator/platform-services-impl/src/main/java/esa/mo/platform/impl/provider/softsim/SoftwareDefinedRadioSoftSimAdapter.java
@@ -22,7 +22,7 @@ package esa.mo.platform.impl.provider.softsim;
 
 import esa.mo.platform.impl.provider.gen.SoftwareDefinedRadioAdapterInterface;
 import opssat.simulator.main.ESASimulator;
-import org.ccsds.moims.mo.mal.structures.DoubleList;
+import org.ccsds.moims.mo.mal.structures.FloatList;
 import org.ccsds.moims.mo.platform.softwaredefinedradio.structures.IQComponents;
 import org.ccsds.moims.mo.platform.softwaredefinedradio.structures.IQComponentsList;
 import org.ccsds.moims.mo.platform.softwaredefinedradio.structures.SDRConfiguration;
@@ -55,12 +55,13 @@ public class SoftwareDefinedRadioSoftSimAdapter implements SoftwareDefinedRadioA
         int nSamples = 10000;
         double[] data = instrumentsSimulator.getpSDR().readFromBuffer(nSamples);
         
-        DoubleList inPhase = new DoubleList();
-        DoubleList quadrature = new DoubleList();
-        
+        FloatList inPhase = new FloatList();
+        FloatList quadrature = new FloatList();
+
+        // Trim the accuracy to match NMF interface
         for(int i = 0; i < nSamples; i++){
-            inPhase.add(data[2*i]);
-            quadrature.add(data[2*i+1]);
+            inPhase.add((float) data[2*i]);
+            quadrature.add((float) data[2*i+1]);
         }
  
         IQComponents iqComponents = new IQComponents();
